address comments. use mktemp

+++ storage/dummy/src/backend.rs
@@ -41,7 +41,7 @@ impl StorageBackend for DummyStorageBackend {
         }))
     }
 
-    fn insert(&self, value: &[u8], _expiry: usize) -> BoxFuture<()> {
+    fn insert(&self, value: &[u8], _expiry: u64) -> BoxFuture<()> {
         let inner = self.inner.clone();
         let key = Self::to_key(&value);
         let value_owned = value.to_owned();
